Move shared code into the common source set
