+++ build.gradle.kts
@@ -39,11 +39,14 @@ kotlin {
             }
         }
 
-        val jvmTest by getting {
+        val commonTest by getting {
             dependencies {
                 implementation(kotlin("test"))
             }
         }
+
+        val jvmMain by getting
+        val jvmTest by getting
     }
 }
