+++ src/commonMain/kotlin/io/github/xn32/json5k/Configuration.kt
