+++ src/commonMain/kotlin/io/github/xn32/json5k/serialization/StructEncoder.kt
