+++ src/commonTest/kotlin/io/github/xn32/json5k/binding/SerializationTest.kt
