+++ src/commonTest/kotlin/io/github/xn32/json5k/internals/FormatGeneratorTest.kt
